third_party/blink/web_tests/platform/generic/media/media-fragments/TC0035-expected.txt


Title: TC0035
Fragment: '%74=3'
Comment: The media is requested from 3 seconds to the end. Note: %74 is equivalent to t.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
